Set up the project for building and created a build script (build.bat)
Added `<PublishSingleFile>` and `<PublishTrimmed>` properties in project file in order to build the app in a single file.

Created the `build.bat` script to run the build command quickly and copy the final exe into an easy to access directory.

Therefore, you can now use the pre-built binary in the `build/release` folder